backend/requirements.txt
# Backend dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.3
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=5.0.0
email-validator>=2.0.0

# Email
resend>=0.7.0

# Existing dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
schedule>=1.2.0

# Optional: Task queue
celery>=5.3.0
redis>=5.0.0

# Development
pytest>=7.4.0
